Plots for AN V1  (#381)
* update onnx export

* use pt weights

* try to solve too many open files

* update plotting
