+++ mlpf/model/training.py
@@ -4,21 +4,14 @@ import pickle as pkl
 import time
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import Optional
 import logging
-import shutil
-from datetime import datetime
 import tqdm
 import yaml
-import csv
 import json
 import sklearn
 import sklearn.metrics
 import numpy as np
-import pandas
-import matplotlib
-import matplotlib.pyplot as plt
-import glob
+import csv
 
 # comet needs to be imported before torch
 from comet_ml import OfflineExperiment, Experiment  # noqa: F401, isort:skip
@@ -26,8 +19,6 @@ from comet_ml import OfflineExperiment, Experiment  # noqa: F401, isort:skip
 import torch
 import torch.distributed as dist
 import torch.multiprocessing as mp
-from torch import Tensor, nn
-from torch.nn import functional as F
 from torch.profiler import ProfilerActivity, profile, record_function
 from torch.utils.tensorboard import SummaryWriter
 
@@ -35,7 +26,6 @@ from mlpf.model.logger import _logger, _configLogger
 from mlpf.model.utils import (
     unpack_predictions,
     unpack_target,
-    get_model_state_dict,
     load_checkpoint,
     save_checkpoint,
     CLASS_LABELS,
@@ -51,211 +41,9 @@ from mlpf.model.inference import make_plots, run_predictions
 from mlpf.model.mlpf import set_save_attention
 from mlpf.model.mlpf import MLPF
 from mlpf.model.PFDataset import Collater, PFDataset, get_interleaved_dataloaders
-
+from mlpf.model.losses import mlpf_loss
 from mlpf.utils import create_comet_experiment
-
-
-def sliced_wasserstein_loss(y_pred, y_true, num_projections=200):
-    # create normalized random basis vectors
-    theta = torch.randn(num_projections, y_true.shape[-1]).to(device=y_true.device)
-    theta = theta / torch.sqrt(torch.sum(theta**2, axis=1, keepdims=True))
-
-    # project the features with the random basis
-    A = torch.matmul(y_true, torch.transpose(theta, -1, -2))
-    B = torch.matmul(y_pred, torch.transpose(theta, -1, -2))
-
-    A_sorted = torch.sort(A, axis=-2).values
-    B_sorted = torch.sort(B, axis=-2).values
-
-    ret = torch.sqrt(torch.sum(torch.pow(A_sorted - B_sorted, 2), axis=[-1, -2]))
-    return ret
-
-
-def mlpf_loss(y, ypred, batch):
-    """
-    Args
-        y [dict]: relevant keys are "cls_id, momentum, charge"
-        ypred [dict]: relevant keys are "cls_id_onehot, momentum, charge"
-        batch [PFBatch]: the MLPF inputs
-    """
-    loss = {}
-    loss_obj_id = FocalLoss(gamma=2.0, reduction="none")
-
-    msk_pred_particle = torch.unsqueeze((ypred["cls_id"] != 0).to(dtype=torch.float32), axis=-1)
-    msk_true_particle = torch.unsqueeze((y["cls_id"] != 0).to(dtype=torch.float32), axis=-1)
-    nelem = torch.sum(batch.mask)
-    npart = torch.sum(y["cls_id"] != 0)
-
-    ypred["momentum"] = ypred["momentum"] * msk_true_particle
-    y["momentum"] = y["momentum"] * msk_true_particle
-
-    # in case of the 3D-padded mode, pytorch expects (batch, num_classes, ...)
-    ypred["cls_binary"] = ypred["cls_binary"].permute((0, 2, 1))
-    ypred["cls_id_onehot"] = ypred["cls_id_onehot"].permute((0, 2, 1))
-
-    # binary loss for particle / no-particle classification
-    # loss_binary_classification = loss_obj_id(ypred["cls_binary"], (y["cls_id"] != 0).long()).reshape(y["cls_id"].shape)
-    loss_binary_classification = 10 * torch.nn.functional.cross_entropy(ypred["cls_binary"], (y["cls_id"] != 0).long(), reduction="none")
-
-    # compare the particle type, only for cases where there was a true particle
-    loss_pid_classification = loss_obj_id(ypred["cls_id_onehot"], y["cls_id"]).reshape(y["cls_id"].shape)
-    loss_pid_classification[y["cls_id"] == 0] *= 0
-
-    # compare particle momentum, only for cases where there was a true particle
-    loss_regression_pt = torch.nn.functional.mse_loss(ypred["pt"], y["pt"], reduction="none")
-    loss_regression_eta = 1e-2 * torch.nn.functional.mse_loss(ypred["eta"], y["eta"], reduction="none")
-    loss_regression_sin_phi = 1e-2 * torch.nn.functional.mse_loss(ypred["sin_phi"], y["sin_phi"], reduction="none")
-    loss_regression_cos_phi = 1e-2 * torch.nn.functional.mse_loss(ypred["cos_phi"], y["cos_phi"], reduction="none")
-    loss_regression_energy = torch.nn.functional.mse_loss(ypred["energy"], y["energy"], reduction="none")
-
-    loss_regression_pt[y["cls_id"] == 0] *= 0
-    loss_regression_eta[y["cls_id"] == 0] *= 0
-    loss_regression_sin_phi[y["cls_id"] == 0] *= 0
-    loss_regression_cos_phi[y["cls_id"] == 0] *= 0
-    loss_regression_energy[y["cls_id"] == 0] *= 0
-
-    # set the loss to 0 on padded elements in the batch
-    loss_binary_classification[batch.mask == 0] *= 0
-    loss_pid_classification[batch.mask == 0] *= 0
-    loss_regression_pt[batch.mask == 0] *= 0
-    loss_regression_eta[batch.mask == 0] *= 0
-    loss_regression_sin_phi[batch.mask == 0] *= 0
-    loss_regression_cos_phi[batch.mask == 0] *= 0
-    loss_regression_energy[batch.mask == 0] *= 0
-
-    # average over all target particles
-    loss["Regression_pt"] = loss_regression_pt.sum() / npart
-    loss["Regression_eta"] = loss_regression_eta.sum() / npart
-    loss["Regression_sin_phi"] = loss_regression_sin_phi.sum() / npart
-    loss["Regression_cos_phi"] = loss_regression_cos_phi.sum() / npart
-    loss["Regression_energy"] = loss_regression_energy.sum() / npart
-
-    # average over all elements that were not padded
-    loss["Classification_binary"] = loss_binary_classification.sum() / nelem
-    loss["Classification"] = loss_pid_classification.sum() / nelem
-
-    # compute predicted pt from model output
-    pred_pt = torch.unsqueeze(torch.exp(ypred["pt"]) * batch.X[..., 1], axis=-1) * msk_pred_particle
-    pred_px = pred_pt * torch.unsqueeze(ypred["cos_phi"].detach(), axis=-1) * msk_pred_particle
-    pred_py = pred_pt * torch.unsqueeze(ypred["sin_phi"].detach(), axis=-1) * msk_pred_particle
-    # pred_pz = pred_pt * torch.unsqueeze(torch.sinh(ypred["eta"].detach()), axis=-1) * msk_pred_particle
-    # pred_mass2 = pred_e**2 - pred_pt**2 - pred_pz**2
-
-    # compute MET, sum across particle axis in event
-    pred_met = torch.sqrt(torch.sum(pred_px, axis=-2) ** 2 + torch.sum(pred_py, axis=-2) ** 2).detach()
-    loss["MET"] = torch.nn.functional.huber_loss(pred_met.squeeze(dim=-1), batch.genmet).mean()
-
-    was_input_pred = torch.concat([torch.softmax(ypred["cls_binary"].transpose(1, 2), axis=-1), ypred["momentum"]], axis=-1) * batch.mask.unsqueeze(
-        axis=-1
-    )
-    was_input_true = torch.concat([torch.nn.functional.one_hot((y["cls_id"] != 0).to(torch.long)), y["momentum"]], axis=-1) * batch.mask.unsqueeze(
-        axis=-1
-    )
-
-    # standardize Wasserstein loss
-    std = was_input_true[batch.mask].std(axis=0)
-    loss["Sliced_Wasserstein_Loss"] = sliced_wasserstein_loss(was_input_pred / std, was_input_true / std).mean()
-
-    # this is the final loss to be optimized
-    loss["Total"] = (
-        loss["Classification_binary"]
-        + loss["Classification"]
-        + loss["Regression_pt"]
-        + loss["Regression_eta"]
-        + loss["Regression_sin_phi"]
-        + loss["Regression_cos_phi"]
-        + loss["Regression_energy"]
-    )
-
-    # store these separately but detached
-    loss["Classification_binary"] = loss["Classification_binary"].detach()
-    loss["Classification"] = loss["Classification"].detach()
-    loss["Regression_pt"] = loss["Regression_pt"].detach()
-    loss["Regression_eta"] = loss["Regression_eta"].detach()
-    loss["Regression_sin_phi"] = loss["Regression_sin_phi"].detach()
-    loss["Regression_cos_phi"] = loss["Regression_cos_phi"].detach()
-    loss["Regression_energy"] = loss["Regression_energy"].detach()
-    loss["Sliced_Wasserstein_Loss"] = loss["Sliced_Wasserstein_Loss"].detach()
-
-    # print out losses for debugging
-    # for k in sorted(loss.keys()):
-    #     print(k, loss[k].item())
-
-    return loss
-
-
-# from https://github.com/AdeelH/pytorch-multi-class-focal-loss/blob/master/focal_loss.py
-class FocalLoss(nn.Module):
-    """Focal Loss, as described in https://arxiv.org/abs/1708.02002.
-    It is essentially an enhancement to cross entropy loss and is
-    useful for classification tasks when there is a large class imbalance.
-    x is expected to contain raw, unnormalized scores for each class.
-    y is expected to contain class labels.
-    Shape:
-        - x: (batch_size, C) or (batch_size, C, d1, d2, ..., dK), K > 0.
-        - y: (batch_size,) or (batch_size, d1, d2, ..., dK), K > 0.
-    """
-
-    def __init__(self, alpha: Optional[Tensor] = None, gamma: float = 0.0, reduction: str = "mean", ignore_index: int = -100):
-        """Constructor.
-        Args:
-            alpha (Tensor, optional): Weights for each class. Defaults to None.
-            gamma (float, optional): A constant, as described in the paper.
-                Defaults to 0.
-            reduction (str, optional): 'mean', 'sum' or 'none'.
-                Defaults to 'mean'.
-            ignore_index (int, optional): class label to ignore.
-                Defaults to -100.
-        """
-        if reduction not in ("mean", "sum", "none"):
-            raise ValueError('Reduction must be one of: "mean", "sum", "none".')
-
-        super().__init__()
-        self.alpha = alpha
-        self.gamma = gamma
-        self.reduction = reduction
-
-        self.nll_loss = nn.NLLLoss(weight=alpha, reduction="none")
-
-    def __repr__(self):
-        arg_keys = ["alpha", "gamma", "reduction"]
-        arg_vals = [self.__dict__[k] for k in arg_keys]
-        arg_strs = [f"{k}={v!r}" for k, v in zip(arg_keys, arg_vals)]
-        arg_str = ", ".join(arg_strs)
-        return f"{type(self).__name__}({arg_str})"
-
-    def forward(self, x: Tensor, y: Tensor) -> Tensor:
-        if x.ndim > 2:
-            # (N, C, d1, d2, ..., dK) --> (N * d1 * ... * dK, C)
-            c = x.shape[1]
-            x = x.permute(0, *range(2, x.ndim), 1).reshape(-1, c)
-            # (N, d1, d2, ..., dK) --> (N * d1 * ... * dK,)
-            y = y.view(-1)
-
-        # compute weighted cross entropy term: -alpha * log(pt)
-        # (alpha is already part of self.nll_loss)
-        log_p = F.log_softmax(x, dim=-1)
-        ce = self.nll_loss(log_p, y)
-
-        # get true class column from each row
-        # this is slow due to indexing
-        # all_rows = torch.arange(len(x))
-        # log_pt = log_p[all_rows, y]
-        log_pt = torch.gather(log_p, 1, y.unsqueeze(axis=-1)).squeeze(axis=-1)
-
-        # compute focal term: (1 - pt)^gamma
-        pt = log_pt.exp()
-        focal_term = (1 - pt) ** self.gamma
-
-        # the full loss: -alpha * ((1 - pt)^gamma) * log(pt)
-        loss = focal_term * ce
-
-        if self.reduction == "mean":
-            loss = loss.mean()
-        elif self.reduction == "sum":
-            loss = loss.sum()
-
-        return loss
+from mlpf.model.plots import validation_plots
 
 
 def configure_model_trainable(model, trainable, is_training):
@@ -277,154 +65,7 @@ def configure_model_trainable(model, trainable, is_training):
         model.eval()
 
 
-def validation_plots(batch, ypred_raw, ytarget, ypred, tensorboard_writer, epoch, outdir):
-    X = batch.X[batch.mask].cpu()
-    ytarget_flat = batch.ytarget[batch.mask].cpu()
-    ypred_binary = ypred_raw[0][batch.mask].detach().cpu()
-    ypred_binary_cls = torch.argmax(ypred_binary, axis=-1)
-    ypred_cls = ypred_raw[1][batch.mask].detach().cpu()
-    ypred_p4 = ypred_raw[2][batch.mask].detach().cpu()
-
-    arr = torch.concatenate(
-        [X, ytarget_flat, ypred_binary, ypred_cls, ypred_p4],
-        axis=-1,
-    ).numpy()
-    df = pandas.DataFrame(arr)
-    df.to_parquet(f"{outdir}/batch0_epoch{epoch}.parquet")
-
-    if tensorboard_writer:
-        sig_prob = torch.softmax(ypred_binary, axis=-1)[:, 1].to(torch.float32)
-        for xcls in np.unique(X[:, 0]):
-            fig = plt.figure()
-            msk = X[:, 0] == xcls
-            etarget = ytarget_flat[msk & (ytarget_flat[:, 0] != 0), 6]
-            epred = ypred_p4[msk & (ypred_binary_cls != 0), 4]
-            b = np.linspace(-2, 2, 100)
-            plt.hist(etarget, bins=b, histtype="step")
-            plt.hist(epred, bins=b, histtype="step")
-            plt.xlabel("log [E/E_elem]")
-            plt.yscale("log")
-            tensorboard_writer.add_figure("energy_elemtype{}".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure()
-            msk = X[:, 0] == xcls
-            pt_target = ytarget_flat[msk & (ytarget_flat[:, 0] != 0), 2]
-            pt_pred = ypred_p4[msk & (ypred_binary_cls != 0), 0]
-            b = np.linspace(-2, 2, 100)
-            plt.hist(etarget, bins=b, histtype="step")
-            plt.hist(epred, bins=b, histtype="step")
-            plt.xlabel("log [pt/pt_elem]")
-            plt.yscale("log")
-            tensorboard_writer.add_figure("pt_elemtype{}".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure(figsize=(5, 5))
-            msk = (X[:, 0] == xcls) & (ytarget_flat[:, 0] != 0) & (ypred_binary_cls != 0)
-            etarget = ytarget_flat[msk, 6]
-            epred = ypred_p4[msk, 4]
-            b = np.linspace(-2, 2, 100)
-            plt.hist2d(etarget, epred, bins=b, cmap="hot", norm=matplotlib.colors.LogNorm())
-            plt.plot([-4, 4], [-4, 4], color="black", ls="--")
-            plt.xlabel("log [E_target/E_elem]")
-            plt.ylabel("log [E_pred/E_elem]")
-            tensorboard_writer.add_figure("energy_elemtype{}_corr".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure(figsize=(5, 5))
-            msk = (X[:, 0] == xcls) & (ytarget_flat[:, 0] != 0) & (ypred_binary_cls != 0)
-            pt_target = ytarget_flat[msk, 2]
-            pt_pred = ypred_p4[msk, 0]
-            b = np.linspace(-2, 2, 100)
-            plt.hist2d(pt_target, pt_pred, bins=b, cmap="hot", norm=matplotlib.colors.LogNorm())
-            plt.plot([-4, 4], [-4, 4], color="black", ls="--")
-            plt.xlabel("log [pt_target/pt_elem]")
-            plt.ylabel("log [pt_pred/pt_elem]")
-            tensorboard_writer.add_figure("pt_elemtype{}_corr".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure(figsize=(5, 5))
-            msk = (X[:, 0] == xcls) & (ytarget_flat[:, 0] != 0) & (ypred_binary_cls != 0)
-            eta_target = ytarget_flat[msk, 3]
-            eta_pred = ypred_p4[msk, 1]
-            b = np.linspace(-6, 6, 100)
-            plt.hist2d(eta_target, eta_pred, bins=b, cmap="hot", norm=matplotlib.colors.LogNorm())
-            plt.plot([-6, 6], [-6, 6], color="black", ls="--")
-            plt.xlabel("eta_target")
-            plt.ylabel("eta_pred")
-            tensorboard_writer.add_figure("eta_elemtype{}_corr".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure(figsize=(5, 5))
-            msk = (X[:, 0] == xcls) & (ytarget_flat[:, 0] != 0) & (ypred_binary_cls != 0)
-            sphi_target = ytarget_flat[msk, 4]
-            sphi_pred = ypred_p4[msk, 2]
-            b = np.linspace(-2, 2, 100)
-            plt.hist2d(sphi_target, sphi_pred, bins=b, cmap="hot", norm=matplotlib.colors.LogNorm())
-            plt.plot([-2, 2], [-2, 2], color="black", ls="--")
-            plt.xlabel("sin_phi_target")
-            plt.ylabel("sin_phi_pred")
-            tensorboard_writer.add_figure("sphi_elemtype{}_corr".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure(figsize=(5, 5))
-            msk = (X[:, 0] == xcls) & (ytarget_flat[:, 0] != 0) & (ypred_binary_cls != 0)
-            cphi_target = ytarget_flat[msk, 5]
-            cphi_pred = ypred_p4[msk, 3]
-            b = np.linspace(-2, 2, 100)
-            plt.hist2d(cphi_target, cphi_pred, bins=b, cmap="hot", norm=matplotlib.colors.LogNorm())
-            plt.plot([-2, 2], [-2, 2], color="black", ls="--")
-            plt.xlabel("cos_phi_target")
-            plt.ylabel("cos_phi_pred")
-            tensorboard_writer.add_figure("cphi_elemtype{}_corr".format(int(xcls)), fig, global_step=epoch)
-
-            fig = plt.figure()
-            msk = X[:, 0] == xcls
-            b = np.linspace(0, 1, 100)
-            plt.hist(sig_prob[msk & (ytarget_flat[:, 0] == 0)], bins=b, histtype="step")
-            plt.hist(sig_prob[msk & (ytarget_flat[:, 0] != 0)], bins=b, histtype="step")
-            plt.xlabel("particle proba")
-            tensorboard_writer.add_figure("sig_proba_elemtype{}".format(int(xcls)), fig, global_step=epoch)
-
-        tensorboard_writer.add_histogram("pt_target", torch.clamp(batch.ytarget[batch.mask][:, 2], -10, 10), global_step=epoch)
-        tensorboard_writer.add_histogram("pt_pred", torch.clamp(ypred_raw[2][batch.mask][:, 0], -10, 10), global_step=epoch)
-        ratio = (ypred_raw[2][batch.mask][:, 0] / batch.ytarget[batch.mask][:, 2])[batch.ytarget[batch.mask][:, 0] != 0]
-        tensorboard_writer.add_histogram("pt_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
-
-        tensorboard_writer.add_histogram("eta_target", torch.clamp(batch.ytarget[batch.mask][:, 3], -10, 10), global_step=epoch)
-        tensorboard_writer.add_histogram("eta_pred", torch.clamp(ypred_raw[2][batch.mask][:, 1], -10, 10), global_step=epoch)
-        ratio = (ypred_raw[2][batch.mask][:, 1] / batch.ytarget[batch.mask][:, 3])[batch.ytarget[batch.mask][:, 0] != 0]
-        tensorboard_writer.add_histogram("eta_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
-
-        tensorboard_writer.add_histogram("sphi_target", torch.clamp(batch.ytarget[batch.mask][:, 4], -10, 10), global_step=epoch)
-        tensorboard_writer.add_histogram("sphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 2], -10, 10), global_step=epoch)
-        ratio = (ypred_raw[2][batch.mask][:, 2] / batch.ytarget[batch.mask][:, 4])[batch.ytarget[batch.mask][:, 0] != 0]
-        tensorboard_writer.add_histogram("sphi_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
-
-        tensorboard_writer.add_histogram("cphi_target", torch.clamp(batch.ytarget[batch.mask][:, 5], -10, 10), global_step=epoch)
-        tensorboard_writer.add_histogram("cphi_pred", torch.clamp(ypred_raw[2][batch.mask][:, 3], -10, 10), global_step=epoch)
-        ratio = (ypred_raw[2][batch.mask][:, 3] / batch.ytarget[batch.mask][:, 5])[batch.ytarget[batch.mask][:, 0] != 0]
-        tensorboard_writer.add_histogram("cphi_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
-
-        tensorboard_writer.add_histogram("energy_target", torch.clamp(batch.ytarget[batch.mask][:, 6], -10, 10), global_step=epoch)
-        tensorboard_writer.add_histogram("energy_pred", torch.clamp(ypred_raw[2][batch.mask][:, 4], -10, 10), global_step=epoch)
-        ratio = (ypred_raw[2][batch.mask][:, 4] / batch.ytarget[batch.mask][:, 6])[batch.ytarget[batch.mask][:, 0] != 0]
-        tensorboard_writer.add_histogram("energy_ratio", torch.clamp(ratio, -10, 10), global_step=epoch)
-
-        for attn in sorted(list(glob.glob(f"{outdir}/attn_conv_*.npz"))):
-            attn_name = os.path.basename(attn).split(".")[0]
-            attn_matrix = np.load(attn)["att"]
-            batch_size = min(attn_matrix.shape[0], 8)
-            fig, axes = plt.subplots(1, batch_size, figsize=((batch_size * 3, 1 * 3)))
-            if isinstance(axes, matplotlib.axes._axes.Axes):
-                axes = [axes]
-            for ibatch in range(batch_size):
-                plt.sca(axes[ibatch])
-                # plot the attention matrix of the first event in the batch
-                plt.imshow(attn_matrix[ibatch].T, cmap="hot", norm=matplotlib.colors.LogNorm())
-                plt.xticks([])
-                plt.yticks([])
-                plt.colorbar()
-                plt.title("event {}, m={:.2E}".format(ibatch, np.mean(attn_matrix[ibatch][attn_matrix[ibatch] > 0])))
-            plt.suptitle(attn_name)
-            tensorboard_writer.add_figure(attn_name, fig, global_step=epoch)
-
-
-def train_and_valid(
+def one_epoch(
     rank,
     world_size,
     outdir,
@@ -442,6 +83,7 @@ def train_and_valid(
     dtype=torch.float32,
     tensorboard_writer=None,
     save_attention=False,
+    checkpoint_dir="",
 ):
     """
     Performs training over a given epoch. Will run a validation step every N_STEPS and after the last training batch.
@@ -541,7 +183,7 @@ def train_and_valid(
                     loss_accum = 0.0
 
                     extra_state = {"step": step, "lr_schedule_state_dict": lr_schedule.state_dict()}
-                    save_checkpoint(f"{outdir}/step_weights.pth", model, optimizer, extra_state)
+                    save_checkpoint(f"{checkpoint_dir}/step_weights.pth", model, optimizer, extra_state)
 
             if not (comet_experiment is None) and (itrain % comet_step_freq == 0):
                 # this loss is not normalized to batch size
@@ -563,7 +205,7 @@ def train_and_valid(
                     intermediate_losses_t[loss_] = intermediate_losses_t[loss_].cpu().item() / itrain
 
                 # compute intermediate validation loss
-                intermediate_losses_v = train_and_valid(
+                intermediate_losses_v = one_epoch(
                     rank,
                     world_size,
                     outdir,
@@ -574,6 +216,7 @@ def train_and_valid(
                     is_train=False,
                     epoch=epoch,
                     dtype=dtype,
+                    checkpoint_dir=checkpoint_dir,
                 )
                 intermediate_metrics = dict(
                     loss=intermediate_losses_t["Total"],
@@ -637,7 +280,7 @@ def train_and_valid(
     return epoch_loss
 
 
-def train_mlpf(
+def train_all_epochs(
     rank,
     world_size,
     model,
@@ -657,9 +300,10 @@ def train_mlpf(
     comet_step_freq=None,
     val_freq=None,
     save_attention=False,
+    checkpoint_dir="",
 ):
     """
-    Will run a full training by calling train().
+    Will run the training pipeline
 
     Args:
         rank: 'cpu' or int representing the gpu device id
@@ -704,7 +348,7 @@ def train_mlpf(
         if epoch == -1:
             with profile(activities=[ProfilerActivity.CPU, ProfilerActivity.CUDA], record_shapes=True, with_stack=True) as prof:
                 with record_function("model_train"):
-                    losses_t = train_and_valid(
+                    losses_t = one_epoch(
                         rank,
                         world_size,
                         outdir,
@@ -717,10 +361,11 @@ def train_mlpf(
                         lr_schedule=lr_schedule,
                         val_freq=val_freq,
                         dtype=dtype,
+                        checkpoint_dir=checkpoint_dir,
                     )
             prof.export_chrome_trace("trace.json")
         else:
-            losses_t = train_and_valid(
+            losses_t = one_epoch(
                 rank,
                 world_size,
                 outdir,
@@ -737,10 +382,11 @@ def train_mlpf(
                 val_freq=val_freq,
                 dtype=dtype,
                 tensorboard_writer=tensorboard_writer_train,
+                checkpoint_dir=checkpoint_dir,
             )
         t_train = time.time()  # epoch time excluding validation
 
-        losses_v = train_and_valid(
+        losses_v = one_epoch(
             rank,
             world_size,
             outdir,
@@ -757,6 +403,7 @@ def train_mlpf(
             dtype=dtype,
             tensorboard_writer=tensorboard_writer_valid,
             save_attention=save_attention,
+            checkpoint_dir=checkpoint_dir,
         )
         t_valid = time.time()
 
@@ -772,17 +419,11 @@ def train_mlpf(
             if losses_v["Total"] < best_val_loss:
                 best_val_loss = losses_v["Total"]
                 stale_epochs = 0
-                torch.save(
-                    {"model_state_dict": get_model_state_dict(model), "optimizer_state_dict": optimizer.state_dict()},
-                    f"{outdir}/best_weights.pth",
-                )
-                save_checkpoint(f"{outdir}/best_weights.pth", model, optimizer, extra_state)
+                save_checkpoint(f"{checkpoint_dir}/best_weights.pth", model, optimizer, extra_state)
             else:
                 stale_epochs += 1
 
             if checkpoint_freq and (epoch != 0) and (epoch % checkpoint_freq == 0):
-                checkpoint_dir = Path(outdir) / "checkpoints"
-                checkpoint_dir.mkdir(exist_ok=True)
                 checkpoint_path = "{}/checkpoint-{:02d}-{:.6f}.pth".format(checkpoint_dir, epoch, losses_v["Total"])
                 save_checkpoint(checkpoint_path, model, optimizer, extra_state)
 
@@ -827,6 +468,7 @@ def train_mlpf(
                 ray.train.report(metrics)
 
         if stale_epochs > patience:
+            logging.info(f"breaking due to stale epochs: {stale_epochs}")
             break
 
         if (rank == 0) or (rank == "cpu"):
@@ -897,6 +539,8 @@ def run(rank, world_size, config, args, outdir, logfile):
         dist.init_process_group("nccl", rank=rank, world_size=world_size)  # (nccl should be faster than gloo)
 
     start_epoch = 1
+    checkpoint_dir = Path(outdir) / "checkpoints"
+    checkpoint_dir.mkdir(parents=True, exist_ok=True)
 
     if config["load"]:  # load a pre-trained model
         with open(f"{outdir}/model_kwargs.pkl", "rb") as f:
@@ -910,7 +554,10 @@ def run(rank, world_size, config, args, outdir, logfile):
         optimizer = torch.optim.AdamW(model.parameters(), lr=config["lr"])
 
         checkpoint = torch.load(config["load"], map_location=torch.device(rank))
-        start_epoch = checkpoint["extra_state"]["epoch"] + 1
+
+        # check if we reached the first epoch in the checkpoint
+        if "epoch" in checkpoint["extra_state"]:
+            start_epoch = checkpoint["extra_state"]["epoch"] + 1
 
         missing_keys, strict = [], True
         for k in model.state_dict().keys():
@@ -1011,7 +658,7 @@ def run(rank, world_size, config, args, outdir, logfile):
         last_epoch = -1 if start_epoch == 1 else start_epoch - 1
         lr_schedule = get_lr_schedule(config, optimizer, config["num_epochs"], steps_per_epoch, last_epoch)
 
-        train_mlpf(
+        train_all_epochs(
             rank,
             world_size,
             model,
@@ -1031,9 +678,10 @@ def run(rank, world_size, config, args, outdir, logfile):
             comet_step_freq=config["comet_step_freq"],
             val_freq=config["val_freq"],
             save_attention=config["save_attention"],
+            checkpoint_dir=checkpoint_dir,
         )
 
-        checkpoint = torch.load(f"{outdir}/best_weights.pth", map_location=torch.device(rank))
+        checkpoint = torch.load(f"{checkpoint_dir}/best_weights.pth", map_location=torch.device(rank))
         model, optimizer = load_checkpoint(checkpoint, model, optimizer)
 
     if not (config["load"] is None):
@@ -1119,9 +767,12 @@ def run(rank, world_size, config, args, outdir, logfile):
 
     if (rank == 0) or (rank == "cpu"):  # make plots only on a single machine
         if args.make_plots:
+
+            ntest_files = -1
+            # ntest_files = 1000
             for sample in args.test_datasets:
                 _logger.info(f"Plotting distributions for {sample}")
-                make_plots(outdir, sample, config["dataset"], testdir_name)
+                make_plots(outdir, sample, config["dataset"], testdir_name, ntest_files)
 
     if world_size > 1:
         dist.destroy_process_group()
@@ -1148,7 +799,6 @@ def override_config(config, args):
 
 
 def device_agnostic_run(config, args, world_size, outdir):
-
     if args.train:
         logfile = f"{outdir}/train.log"
     else:
@@ -1181,304 +831,3 @@ def device_agnostic_run(config, args, world_size, outdir):
         rank = "cpu"
         _logger.info("Will use cpu", color="purple")
         run(rank, world_size, config, args, outdir, logfile)
-
-
-def train_ray_trial(config, args, outdir=None):
-    import ray
-
-    if outdir is None:
-        outdir = ray.train.get_context().get_trial_dir()
-        if not os.path.exists(outdir):
-            os.makedirs(outdir)
-
-    use_cuda = args.gpus > 0
-
-    rank = ray.train.get_context().get_local_rank() if use_cuda else "cpu"
-    world_rank = ray.train.get_context().get_world_rank()
-    world_size = ray.train.get_context().get_world_size()
-
-    model_kwargs = {
-        "input_dim": len(X_FEATURES[config["dataset"]]),
-        "num_classes": len(CLASS_LABELS[config["dataset"]]),
-        "input_encoding": config["model"]["input_encoding"],
-        "pt_mode": config["model"]["pt_mode"],
-        "eta_mode": config["model"]["eta_mode"],
-        "sin_phi_mode": config["model"]["sin_phi_mode"],
-        "cos_phi_mode": config["model"]["cos_phi_mode"],
-        "energy_mode": config["model"]["energy_mode"],
-        "elemtypes_nonzero": ELEM_TYPES_NONZERO[config["dataset"]],
-        "learned_representation_mode": config["model"]["learned_representation_mode"],
-        **config["model"][config["conv_type"]],
-    }
-    model = MLPF(**model_kwargs)
-
-    if world_size > 1:
-        model = torch.nn.SyncBatchNorm.convert_sync_batchnorm(model)
-    # optimizer should be created after distributing the model to devices with ray.train.torch.prepare_model(model)
-    model = ray.train.torch.prepare_model(model)
-    optimizer = torch.optim.AdamW(model.parameters(), lr=config["lr"])
-
-    trainable_params, nontrainable_params, table = count_parameters(model)
-    print(table)
-
-    if (rank == 0) or (rank == "cpu"):
-        with open(os.path.join(outdir, "num_params.csv"), "w", newline="") as f:
-            writer = csv.writer(f)
-            writer.writerow(["trainable_params", "nontrainable_params", "total_params"])
-            writer.writerow([trainable_params, nontrainable_params, trainable_params + nontrainable_params])
-        _logger.info(model)
-        _logger.info(f"Trainable parameters: {trainable_params}")
-        _logger.info(f"Non-trainable parameters: {nontrainable_params}")
-        _logger.info(f"Total parameters: {trainable_params + nontrainable_params}")
-        _logger.info(table)
-
-    if (rank == 0) or (rank == "cpu"):
-        save_HPs(args, model, model_kwargs, outdir)  # save model_kwargs and hyperparameters
-        _logger.info("Creating experiment dir {}".format(outdir))
-        _logger.info(f"Model directory {outdir}", color="bold")
-
-    loaders = get_interleaved_dataloaders(world_size, rank, config, use_cuda, use_ray=True)
-
-    if args.comet:
-        comet_experiment = create_comet_experiment(config["comet_name"], comet_offline=config["comet_offline"], outdir=outdir)
-        comet_experiment.set_name(f"world_rank_{world_rank}_{Path(outdir).name}")
-        comet_experiment.log_parameter("run_id", Path(outdir).name)
-        comet_experiment.log_parameter("world_size", world_size)
-        comet_experiment.log_parameter("rank", rank)
-        comet_experiment.log_parameter("world_rank", world_rank)
-        comet_experiment.log_parameters(config, prefix="config:")
-        comet_experiment.set_model_graph(model)
-        comet_experiment.log_parameter(trainable_params, "trainable_params")
-        comet_experiment.log_parameter(nontrainable_params, "nontrainable_params")
-        comet_experiment.log_parameter(trainable_params + nontrainable_params, "total_trainable_params")
-        comet_experiment.log_code(str(Path(outdir).parent.parent / "mlpf/model/training.py"))
-        comet_experiment.log_code(str(Path(outdir).parent.parent / "mlpf/model/mlpf.py"))
-        comet_experiment.log_code(str(Path(outdir).parent.parent / "mlpf/model/utils.py"))
-        comet_experiment.log_code(str(Path(outdir).parent.parent / "mlpf/pipeline.py"))
-        comet_experiment.log_code(str(Path(outdir).parent.parent / "mlpf/raytune/pt_search_space.py"))
-        # save overridden config then log to comet
-        config_filename = "overridden_config.yaml"
-        with open((Path(outdir) / config_filename), "w") as file:
-            yaml.dump(config, file)
-        comet_experiment.log_code(str(Path(outdir) / config_filename))
-    else:
-        comet_experiment = None
-
-    steps_per_epoch = len(loaders["train"])
-    start_epoch = 1
-    lr_schedule = get_lr_schedule(config, optimizer, config["num_epochs"], steps_per_epoch, last_epoch=-1)
-
-    checkpoint = ray.train.get_checkpoint()
-    if checkpoint:
-        with checkpoint.as_directory() as checkpoint_dir:
-            with checkpoint.as_directory() as checkpoint_dir:
-                checkpoint = torch.load(Path(checkpoint_dir) / "checkpoint.pth", map_location=torch.device(rank))
-                if args.resume_training:
-                    model, optimizer = load_checkpoint(checkpoint, model, optimizer)
-                    start_epoch = checkpoint["extra_state"]["epoch"] + 1
-                    lr_schedule = get_lr_schedule(config, optimizer, config["num_epochs"], steps_per_epoch, last_epoch=start_epoch - 1)
-                else:  # start a new training with model weights loaded from a pre-trained model
-                    model = load_checkpoint(checkpoint, model)
-
-    train_mlpf(
-        rank,
-        world_size,
-        model,
-        optimizer,
-        loaders["train"],
-        loaders["valid"],
-        config["num_epochs"],
-        config["patience"],
-        outdir,
-        trainable=config["model"]["trainable"],
-        start_epoch=start_epoch,
-        lr_schedule=lr_schedule,
-        use_ray=True,
-        checkpoint_freq=config["checkpoint_freq"],
-        comet_experiment=comet_experiment,
-        comet_step_freq=config["comet_step_freq"],
-        dtype=getattr(torch, config["dtype"]),
-        val_freq=config["val_freq"],
-    )
-
-
-def run_ray_training(config, args, outdir):
-    import ray
-    from ray import tune
-    from ray.train.torch import TorchTrainer
-
-    if not args.local:
-        ray.init(address="auto")
-
-    if args.resume_training:
-        outdir = args.resume_training  # continue training in the same directory
-
-    _configLogger("mlpf", filename=f"{outdir}/train.log")
-
-    use_gpu = args.gpus > 0
-    num_workers = args.gpus if use_gpu else 1
-    scaling_config = ray.train.ScalingConfig(
-        num_workers=num_workers,
-        use_gpu=use_gpu,
-        resources_per_worker={"CPU": max(1, args.ray_cpus // num_workers - 1), "GPU": int(use_gpu)},  # -1 to avoid blocking
-    )
-    storage_path = Path(args.experiments_dir if args.experiments_dir else "experiments").resolve()
-    run_config = ray.train.RunConfig(
-        name=Path(outdir).name,
-        storage_path=storage_path,
-        log_to_file=False,
-        failure_config=ray.train.FailureConfig(max_failures=2),
-        checkpoint_config=ray.train.CheckpointConfig(num_to_keep=1),  # keep only latest checkpoint
-        sync_config=ray.train.SyncConfig(sync_artifacts=True),
-    )
-    trainable = tune.with_parameters(train_ray_trial, args=args, outdir=outdir)
-    # Resume from checkpoint if a checkpoitn is found in outdir
-    if TorchTrainer.can_restore(outdir):
-        _logger.info(f"Restoring Ray Trainer from {outdir}", color="bold")
-        trainer = TorchTrainer.restore(outdir, train_loop_per_worker=trainable, scaling_config=scaling_config)
-    else:
-        resume_from_checkpoint = ray.train.Checkpoint(config["load"]) if config["load"] else None
-        if resume_from_checkpoint:
-            _logger.info("Loading checkpoint {}".format(config["load"]), color="bold")
-        trainer = TorchTrainer(
-            train_loop_per_worker=trainable,
-            train_loop_config=config,
-            scaling_config=scaling_config,
-            run_config=run_config,
-            resume_from_checkpoint=resume_from_checkpoint,
-        )
-    result = trainer.fit()
-
-    _logger.info("Final loss: {}".format(result.metrics["loss"]), color="bold")
-    _logger.info("Final cls_loss: {}".format(result.metrics["cls_loss"]), color="bold")
-    _logger.info("Final reg_pt_loss: {}".format(result.metrics["reg_pt_loss"]), color="bold")
-    # _logger.info("Final charge_loss: {}".format(result.metrics["charge_loss"]), color="bold")
-
-    _logger.info("Final val_loss: {}".format(result.metrics["val_loss"]), color="bold")
-    _logger.info("Final val_cls_loss: {}".format(result.metrics["val_cls_loss"]), color="bold")
-    _logger.info("Final val_reg_pt_loss: {}".format(result.metrics["val_reg_pt_loss"]), color="bold")
-    # _logger.info("Final val_charge_loss: {}".format(result.metrics["val_charge_loss"]), color="bold")
-
-
-def set_searchspace_and_run_trial(search_space, config, args):
-    import ray
-    from raytune.pt_search_space import set_hps_from_search_space
-
-    rank = ray.train.get_context().get_local_rank()
-
-    config = set_hps_from_search_space(search_space, config)
-    try:
-        # outdir will be taken from the ray.train.context.TrainContext in each trial
-        train_ray_trial(config, args, outdir=None)
-    except torch.cuda.OutOfMemoryError:
-        ray.train.report({"val_loss": np.NAN})
-        torch.cuda.empty_cache()  # make sure GPU memory is cleared for next trial
-        if rank == 0:
-            logging.warning("OOM error encountered, skipping this hyperparameter configuration.")
-            skiplog_file_path = Path(config["raytune"]["local_dir"]) / args.hpo / "skipped_configurations.txt"
-            lines = ["{}: {}\n".format(item[0], item[1]) for item in search_space.items()]
-
-            with open(skiplog_file_path, "a") as f:
-                f.write("#" * 80 + "\n")
-                for line in lines:
-                    f.write(line)
-                    logging.warning(line[:-1])
-                f.write("#" * 80 + "\n\n")
-            logging.warning("Done writing warnings to log.")
-
-
-def run_hpo(config, args):
-    import ray
-    from ray import tune
-    from ray.train.torch import TorchTrainer
-
-    from raytune.pt_search_space import raytune_num_samples, search_space
-    from raytune.utils import get_raytune_schedule, get_raytune_search_alg
-
-    if args.raytune_num_samples:
-        raytune_num_samples = args.raytune_num_samples  # noqa: F811
-
-    name = args.hpo  # name of Ray Tune experiment directory
-
-    os.environ["TUNE_DISABLE_STRICT_METRIC_CHECKING"] = "1"  # don't crash if a metric is missing
-    if isinstance(config["raytune"]["local_dir"], type(None)):
-        raise TypeError("Please specify a local_dir in the raytune section of the config file.")
-
-    expdir = Path(config["raytune"]["local_dir"]) / name
-    expdir.mkdir(parents=True, exist_ok=True)
-    dirname = Path(config["raytune"]["local_dir"]) / name
-    shutil.copy(
-        "mlpf/raytune/pt_search_space.py",
-        str(dirname / "pt_search_space.py"),
-    )  # Copy the search space definition file to the train dir for later reference
-    # Save config for later reference. Note that saving happens after parameters are overwritten by cmd line args.
-    with open((dirname / "config.yaml"), "w") as file:
-        yaml.dump(config, file)
-
-    if not args.local:
-        _logger.info("Inititalizing ray...")
-        ray.init(
-            address=os.environ["ip_head"],
-            _node_ip_address=os.environ["head_node_ip"],
-        )
-        _logger.info("Done.")
-
-    sched = get_raytune_schedule(config["raytune"])
-    search_alg = get_raytune_search_alg(config["raytune"])
-
-    scaling_config = ray.train.ScalingConfig(
-        num_workers=args.gpus,
-        use_gpu=True,
-        resources_per_worker={"CPU": args.ray_cpus // (args.gpus) - 1, "GPU": 1},  # -1 to avoid blocking
-    )
-
-    if tune.Tuner.can_restore(str(expdir)):
-        args.resume_training = True
-
-    trainable = tune.with_parameters(set_searchspace_and_run_trial, config=config, args=args)
-    trainer = TorchTrainer(train_loop_per_worker=trainable, scaling_config=scaling_config)
-
-    if tune.Tuner.can_restore(str(expdir)):
-        # resume unfinished HPO run
-        tuner = tune.Tuner.restore(str(expdir), trainable=trainer, resume_errored=True, restart_errored=False, resume_unfinished=True)
-    else:
-        # start new HPO run
-        search_space = {"train_loop_config": search_space}  # the ray TorchTrainer only takes a single arg: train_loop_config
-        tuner = tune.Tuner(
-            trainer,
-            param_space=search_space,
-            tune_config=tune.TuneConfig(
-                num_samples=raytune_num_samples,
-                metric=config["raytune"]["default_metric"] if (search_alg is None and sched is None) else None,
-                mode=config["raytune"]["default_mode"] if (search_alg is None and sched is None) else None,
-                search_alg=search_alg,
-                scheduler=sched,
-            ),
-            run_config=ray.train.RunConfig(
-                name=name,
-                storage_path=config["raytune"]["local_dir"],
-                log_to_file=False,
-                failure_config=ray.train.FailureConfig(max_failures=2),
-                checkpoint_config=ray.train.CheckpointConfig(num_to_keep=1),  # keep only latest checkpoint
-                sync_config=ray.train.SyncConfig(sync_artifacts=True),
-            ),
-        )
-    start = datetime.now()
-    _logger.info("Starting tuner.fit()")
-    result_grid = tuner.fit()
-    end = datetime.now()
-
-    print("Number of errored trials: {}".format(result_grid.num_errors))
-    print("Number of terminated (not errored) trials: {}".format(result_grid.num_terminated))
-    print("Ray Tune experiment path: {}".format(result_grid.experiment_path))
-
-    best_result = result_grid.get_best_result(scope="last-10-avg", metric="val_loss", mode="min")
-    best_config = best_result.config
-    print("Best trial path: {}".format(best_result.path))
-
-    result_df = result_grid.get_dataframe()
-    print(result_df)
-    print(result_df.columns)
-
-    logging.info("Total time of Tuner.fit(): {}".format(end - start))
-    logging.info("Best hyperparameters found according to {} were: {}".format(config["raytune"]["default_metric"], best_config))
