+++ scripts/tallinn/a100/pytorch.sh
@@ -4,14 +4,14 @@
 #SBATCH --mem-per-gpu 250G
 #SBATCH -o logs/slurm-%x-%j-%N.out
 
-IMG=/home/software/singularity/pytorch.simg:2024-08-18
+IMG=/home/software/singularity/pytorch.simg:2024-12-03
 cd ~/particleflow
 
-ulimit -n 10000
+ulimit -n 100000
 singularity exec -B /scratch/persistent --nv \
     --env PYTHONPATH=`pwd` \
     --env KERAS_BACKEND=torch \
     $IMG python3 mlpf/pipeline.py --gpus 1 \
-    --data-dir /scratch/persistent/joosep/tensorflow_datasets --config parameters/pytorch/pyg-cms-nopu.yaml \
-    --train --test --make-plots --conv-type attention \
-    --gpu-batch-multiplier 5 --checkpoint-freq 1 --num-workers 8 --prefetch-factor 50 --comet --num-epochs 30
+    --data-dir /scratch/persistent/joosep/tensorflow_datasets --config parameters/pytorch/pyg-cms.yaml \
+    --train --conv-type attention \
+    --gpu-batch-multiplier 5 --checkpoint-freq 1 --num-workers 8 --prefetch-factor 50 --comet
