+++ scripts/tallinn/deploy_cms.sh
@@ -3,7 +3,7 @@
 
 rm -f files_to_copy.txt
 maxfiles=100
-path=experiments/pyg-cms_20241101_090645_682892/./preds_checkpoint-08-2.986092
+path=/local/joosep/mlpf/results/cms/pyg-cms_20241101_090645_682892/./preds_checkpoint-18-2.778778
 targetpath=/home/joosep/huggingface/particleflow/cms/v2.1.0/pyg-cms_20241101_090645_682892/
 
 samplestocopy=(
@@ -19,4 +19,4 @@ for sample in "${samplestocopy[@]}"; do
     find "$path/$sample" -type f | sort | head -n$maxfiles >> files_to_copy.txt
 done
 
-rsync --progress --relative --files-from files_to_copy.txt `pwd` $targetpath
+rsync --progress --relative --files-from files_to_copy.txt / $targetpath
